! Pedantry in action. Making sure all new files have 'LF' svn property. ! messageIconChecks_disable should be enable [bug 4095] ! MySQL 4.0.0 allows for multi-table deletes. (ManageBans.php, ManageMaintenance.php, ManagePermissions.php) ! Two very tiny doc changes that were annoying me + Allow users to view their login history + 1/3 of the way done adding Slammeddime's additional hooks mod ! Always show boxes to add new censored words [bug 4555] (Admin.template.php) ! Robots shouldn't update num views (Display.php) + Style the <body> by id ! getMembersOnlineStats() should be cached and allow for no-sorting as an option (BoardIndex.php, Subs-MembersOnline.php, SSI.php) ! It also appears I fixed a bug where = was used instead of == in Subs-MembersOnline.php + LiteSpeed = Apache [bug 4538] (Querystring.php, Subs.php, Load.php)

+++ Sources/ManagePermissions.php
@@ -996,14 +996,7 @@ function GeneralPermissionSettings($return_config = false)
 		if (empty($modSettings['permission_enable_deny']))
 		{
 			$smcFunc['db_query']('', '
-				DELETE FROM {db_prefix}permissions
-				WHERE add_deny = {int:denied}',
-				array(
-					'denied' => 0,
-				)
-			);
-			$smcFunc['db_query']('', '
-				DELETE FROM {db_prefix}board_permissions
+				DELETE FROM {db_prefix}permissions, {db_prefix}board_permissions
 				WHERE add_deny = {int:denied}',
 				array(
 					'denied' => 0,
@@ -1030,14 +1023,7 @@ function GeneralPermissionSettings($return_config = false)
 
 			// Remove'em.
 			$smcFunc['db_query']('', '
-				DELETE FROM {db_prefix}permissions
-				WHERE id_group IN ({array_int:post_group_list})',
-				array(
-					'post_group_list' => $post_groups,
-				)
-			);
-			$smcFunc['db_query']('', '
-				DELETE FROM {db_prefix}board_permissions
+				DELETE FROM {db_prefix}permissions, {db_prefix}board_permissions
 				WHERE id_group IN ({array_int:post_group_list})',
 				array(
 					'post_group_list' => $post_groups,
